rename token_stream to tokenizer
